Feature #3 Create a Firebase backend and a Firestore database
This sets up a Firebase backend and a Firestore database.

The page is hosted on: https://ci-server-assignment2.web.app/
Here it is explained how to add entries to the database: https://firebase.google.com/docs/firestore/quickstart#add_data

In order to be able to deploy changes to the page, first make sure that you have npm installed on your computer, and that your email address has been added to the project.
Then, run the following commands:

```npm install -g firebase-tools```

```firebase login```

```firebase init```

When initializing, select the existing project called "CI server assignment2". Also, make sure that you only select the two features Firestore and Hosting.
Choose the default filenames when it is asked.

After having made some changes and when it is time to deploy, run the following command:

```firebase deploy```

The changes should then be visible on the page and on Firestore if such changes has been made.

Co-authored-by: Anton Lövström <anton.lovstrom4@gmail.com>