Feature #7 Make a server respond to a github webhook.
This commit shows how to make a server that can recognize
a github webhook and see wich commit has been posted. The
webhook needs to be in json and the URL needs to be the same
as the forwarded ngrok URL. The webhook should only trigger
on push events.

Co-authored-by Olof Gren <olofthegren@gmail.com>