Feature #7 Make a server respond to a github webhook.
This commit shows how to make a server that can recognize
a github webhook and see wich commit has been posted. The
webhook needs to be in json and the URL needs to be the same
as the forwarded ngrok URL. The webhook should only trigger
on push events.

To compile server. Go to dir CI and type:

	$ mvn clean compile assembly:single

Then to run server type:
	$ java -cp target/CI-1.0-SNAPSHOT-jar-with-dependencies.jar group1.Main

To make the localhost accesseble from remote and fore github run ngrok.
To run ngrok type:
	$ ./ngrok http 8080

Co-authored-by Olof Gren <olofthegren@gmail.com>