Feature #13 Email notification functionality
This commit add the email functionality. The pom file have been
updated to include the new dependencies and a file named
email.java have been created. The functionality of email.java is that it can send
emails to any email adress from a newly created email adress.

This functionality is yet to be connected to the CI server.

Co-authored-by: Robin Wänlund <robin.wanlund@hotmail.com>